requirements.txt
Keras==2.2.4
matplotlib==3.1.3
tensorflow==1.13.1
numpy==1.19.5
scikit_image==0.16.2
Pillow==8.3.1
scikit_learn==0.24.2
skimage==0.0
